changes: moved services to use case, to be separated to use case after
