--- backend/src/modules/business/business.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { BusinessService } from './business.service';
-
-describe('BusinessService', () => {
-  let service: BusinessService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [BusinessService],
-    }).compile();
-
-    service = module.get<BusinessService>(BusinessService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
